feat: move Storybook to dedicated app (apps/pops-storybook)
Extracts Storybook configuration from packages/ui into its own app
at apps/pops-storybook, converting main.js to TypeScript. Stories
remain co-located with their components in packages/ui, app-finance,
app-media, and app-inventory — only the dev server and config moved.

Closes #724
