+++ packages/ui/package.json
@@ -15,9 +15,7 @@
     "lint": "eslint src --ext .ts,.tsx",
     "lint:fix": "eslint src --ext .ts,.tsx --fix",
     "format:check": "prettier --check \"src/**/*.{ts,tsx,css}\"",
-    "format:fix": "prettier --write \"src/**/*.{ts,tsx,css}\"",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
+    "format:fix": "prettier --write \"src/**/*.{ts,tsx,css}\""
   },
   "dependencies": {
     "@fontsource-variable/plus-jakarta-sans": "^5.2.8",
@@ -56,9 +54,7 @@
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^5.0.2",
     "@eslint/js": "^9.0.0",
-    "@storybook/addon-a11y": "^10.3.1",
     "@storybook/react-vite": "^10.3.1",
     "@tailwindcss/vite": "4.2.2",
     "@types/react": "^19.2.14",
@@ -66,9 +62,7 @@
     "eslint": "^9.0.0",
     "eslint-plugin-react": "^7.37.0",
     "eslint-plugin-react-hooks": "^7.0.0",
-    "eslint-plugin-storybook": "^0.12.0",
     "prettier": "^3.0.0",
-    "storybook": "^10.3.1",
     "typescript": "^5.7.0",
     "typescript-eslint": "^8.0.0",
     "vite": "^8.0.1"
